move all routes from cli
